precision issue fix.

+++ BeingAliveLanguageCppPort/cppFunc.cpp
@@ -1,20 +1,20 @@
 #include "stdafx.h"
 #include "cppFunc.h"
 
-void BAL_possionDiskElimSample(ON_SimpleArray<float>* inPt, int n,
+void BAL_possionDiskElimSample(ON_SimpleArray<double>* inPt, int n,
 	ON_3dPointArray* outPt) {
 	// input conversion
 	int sz = inPt->Count() / 3;
 
-	std::vector<cy::Vec3f> inputPoints(0);
+	std::vector<cy::Vec3d> inputPoints(0);
 	for (size_t i = 0; i < sz; i++) {
 		inputPoints.emplace_back(*inPt->At(i * 3), *inPt->At(i * 3 + 1),
 			*inPt->At(i * 3 + 2));
 	}
 
 	// elimination to the given number of pts
-	cy::WeightedSampleElimination<cy::Vec3f, float, 2> wse;
-	std::vector<cy::Vec3f> outputPoints(n);
+	cy::WeightedSampleElimination<cy::Vec3d, double, 2> wse;
+	std::vector<cy::Vec3d> outputPoints(n);
 	wse.Eliminate(inputPoints.data(), inputPoints.size(), outputPoints.data(),
 		outputPoints.size(), true);
